Removed intermediate step when signing out. (#32)
Remove the intermediate step when clicking the sign-out link. Now we are
signing out immediately. Also made use of ES6 string interpolation in the
AuthService.

After reading up on
https://facebook.github.io/react/docs/reusable-components.html#no-autobinding
I did the binding for event handlers in the constructor instead of in the
JSX. This hardly matters for us, but I thought it made sense to keep the
JSX as clean as possible and to follow React recommendation.

This closes #7